events initialized

+++ server/src/models/Event.model.js
@@ -37,10 +37,16 @@ const Event = new mongoose.Schema({
         type: Date,
         required: [true, "event must have date"]
     },
-    manager: {
+    managerId: {
         type: mongoose.Schema.Types.ObjectId,
         ref: "user",
-    }
+    },
+    registeredUsers: [
+        {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: "user"
+        }
+    ]
 })
 
 Event.statics.validateEvent = function (title, description, category, poster, duration, eventDate) {
@@ -53,7 +59,7 @@ Event.statics.validateEvent = function (title, description, category, poster, du
     else return true
 }
 
-Event.statics.createEvent = async function (title, description, category, poster, duration, eventDate, tags) {
+Event.statics.createEvent = async function (title, description, category, poster, duration, eventDate, tags, managerId) {
     try {
         await this.create({
             title,
@@ -63,6 +69,7 @@ Event.statics.createEvent = async function (title, description, category, poster
             duration,
             eventDate,
             tags,
+            managerId
         })
     } catch (err) {
         throw new APIError(402, err)
@@ -73,8 +80,11 @@ Event.statics.getEvent = async function (eventId) {
     return await this.findById(eventId)
 }
 
-Event.statics.deleteEvent= async function (eventId) {
+Event.statics.deleteEvent = async function (eventId) {
     try {
+        let event = await this.findById(eventId, managerId)
+        if (!event) throw new APIError(404, "event not found")
+        if (event.managerId !== managerId) throw new APIError(401, "you are not authorized to delete this event")
         await this.findByIdAndDelete(eventId)
     } catch (err) {
         throw new APIError(404, "event not found or issue")
@@ -89,6 +99,18 @@ Event.methods.updateEvent = async function (title, description, category) {
 
         await this.save()
         return true
+    } catch (err) {
+        throw new APIError(402, "event can't be updated")
+    }
+}
+
+Event.methods.joinEvent = async function (userId) {
+    if (!userId) {
+        throw new APIError(402, "userId can not be empty")
+    }
+    try {
+        this.registeredUsers.push(userId)
+        await this.save()
     } catch (err) {
         throw new APIError(402, "event can't be updated")
     }
